Fix namespace lifecycle
Namespaces used to be deleted when they were not in-use anymore, causing
swift deletions and recreations of resources to result in the target
namespace be in deleting state. In of itself that wouldn't be bad,
however, if resources below could not be deleted, the namespace would
not end up being recreated in a sensible timeframe.

Now, namespaces are created when required and only deleted if the source
namespace gets deleted, tieing in better with the deletion flow.