Enhance usage system to avoid unnecessary refs (#24)
Previously, namespaces got a finalizer even if no machine / volume was
scheduled on the pool. This boiled down to the `Dependent` function not
being as modular to filter out e.g. machines that were *not* running on
the pool sometimes.

This change addresses this issue by implementing a revamped dependency
system and using it among all controllers.